audio-study-assistant/requirements.txt
fastapi
uvicorn
python-multipart
google-cloud-storage
google-cloud-aiplatform
